add Exception-Handling instead of Validation, add GET DashboardArticle Route, refactor code, delete access modifier

+++ Application/Articles/Commands/Delete/DeleteArticleCommandHandler.cs
@@ -4,7 +4,9 @@ using System.Threading.Tasks;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using MyFridgeListWebapi.Core.Data.Database;
+using MyFridgeListWebapi.Core.Exceptions;
 using MyFridgeListWebapi.Core.Models.Responses.Article;
+using MyFridgeListWebapi.Properties;
 
 namespace MyFridgeListWebapi.Application.Articles.Commands.Delete
 {
@@ -19,9 +21,23 @@ namespace MyFridgeListWebapi.Application.Articles.Commands.Delete
 
         public async Task<DeleteArticleResponse> Handle(DeleteArticleCommand request, CancellationToken cancellationToken)
         {
+            var fridge = await _dbContext.Fridges
+                .Where(x => x.UserId == request.UserId)
+                .FirstOrDefaultAsync(x => x.Id == request.FridgeId, cancellationToken: cancellationToken);
+
+            if (fridge == null)
+            {
+                throw new NotFoundException(string.Format(Resources.ValidationErrorFridgeWithIdNotExists, request.FridgeId));
+            }
+
             var article = await _dbContext.Articles
-                .Where(x => x.Id == request.ArticleId)
-                .FirstOrDefaultAsync();
+                .Where(x => x.FridgeId == request.FridgeId)
+                .FirstOrDefaultAsync(x => x.Id == request.ArticleId, cancellationToken: cancellationToken);
+
+            if (article == null)
+            {
+                throw new NotFoundException(string.Format(Resources.ValidationErrorArticleWithIdNotExists, request.ArticleId));
+            }
 
             _dbContext.Articles.Remove(article);
             await _dbContext.SaveChangesAsync(cancellationToken);
